rpm: find-legacy-runtime.sh: Use .L suffix for better upgrades.
This makes the legacy package provided by the script always newer
than the original. This improves identification and, together
with a more strict Obsoltes, will cause this new legacy package
to replace some older packages that would not be replaced
otherwise (this also covers a possible case when the new legacy
package happens to have the exact same name as the original one).